// packages/cli/src/enum.ts
export enum CodeLintTools {
  ESLINT_DEFAULT = 'eslint-default',
  ESLINT_NO_EXTERNAL = 'eslint-no-external',
  ESLINT_NO_FORMATTER = 'eslint-no-formatter',
  ESLINT_PRETTIER = 'eslint-prettier',
}

export enum GitLintTools {
  COMMITLINT = 'commitlint',
  CZG = 'czg',
  LINT_STAGED = 'lint-staged',
  HUSKY = 'husky',
}